server/requirements.txt
Django==1.10
psycopg2-binary==2.8.5
django-environ==0.10.0
djangorestframework==3.8.2